# openapi.yaml
openapi: 3.1.0
info:
  title: Multi-Character Drama Engine API
  version: 1.1.0
  description: |
    Runtime-aligned OpenAPI for the current drama stack implementation.
servers:
  - url: /api/v1
...
paths:
...
  /drama/relationships:
    get:
      tags: [drama-relationships]
      summary: List relationships by project or character
      parameters:
        - in: query
          name: project_id
          required: false
          schema:
            type: string
            format: uuid
        - in: query
          name: character_id
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Relationship list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RelationshipRead"
    post:
      tags: [drama-relationships]
      summary: Create relationship
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RelationshipCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelationshipRead"
  /drama/relationships/{relationship_id}:
    patch:
      tags: [drama-relationships]
      summary: Update relationship
      parameters:
        - in: path
          name: relationship_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RelationshipUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelationshipRead"
  /drama/relationships/graph:
    get:
      tags: [drama-relationships]
      summary: Get relationship graph
      parameters:
        - in: query
          name: project_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Graph payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelationshipGraph"
  /drama/relationships/graph/rebuild:
    post:
      tags: [drama-relationships]
      summary: Rebuild relationship graph (current impl returns immediate stub status)
      parameters:
        - in: query
          name: project_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Rebuild response
          content:
            application/json:
              schema:
                type: object
...
  /drama/compile/scene:
    post:
      tags: [drama-compile]
      summary: Compile scene package from analysis payload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompileSceneDirectRequest"
      responses:
        "200":
          description: Compiled scene payload
          content:
            application/json:
              schema:
                type: object
  /drama/compile/episode:
    post:
      tags: [drama-compile]
      summary: Compile episode package from persisted scene rows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompileEpisodeRequest"
      responses:
        "200":
          description: Episode compile response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompileEpisodeResponse"
  /drama/compile/project:
    post:
      tags: [drama-compile]
      summary: Compile project package from persisted scene rows
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompileProjectRequest"
      responses:
        "200":
          description: Project compile response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompileProjectResponse"
...
components:
...
  schemas:
...
    RelationshipCreate:
      type: object
      required: [project_id, source_character_id, target_character_id, relation_type]
      properties:
        project_id: {type: string}
        source_character_id: {type: string, format: uuid}
        target_character_id: {type: string, format: uuid}
        relation_type: {type: string}
        intimacy_level: {type: number}
        trust_level: {type: number}
        dependence_level: {type: number}
        fear_level: {type: number}
        resentment_level: {type: number}
        attraction_level: {type: number}
        rivalry_level: {type: number}
        dominance_source_over_target: {type: number}
        perceived_loyalty: {type: number}
        hidden_agenda_score: {type: number}
        recent_betrayal_score: {type: number}
        unresolved_tension_score: {type: number}
        status: {type: string}
        last_interaction_scene_id: {type: string, nullable: true}
    RelationshipUpdate:
      type: object
      properties:
        relation_type: {type: string}
        intimacy_level: {type: number}
        trust_level: {type: number}
        dependence_level: {type: number}
        fear_level: {type: number}
        resentment_level: {type: number}
        attraction_level: {type: number}
        rivalry_level: {type: number}
        dominance_source_over_target: {type: number}
        perceived_loyalty: {type: number}
        hidden_agenda_score: {type: number}
        recent_betrayal_score: {type: number}
        unresolved_tension_score: {type: number}
        status: {type: string}
        last_interaction_scene_id: {type: string, nullable: true}
    RelationshipRead:
      type: object
      properties:
        id: {type: string, format: uuid}
        project_id: {type: string}
        source_character_id: {type: string, format: uuid}
        target_character_id: {type: string, format: uuid}
        relation_type: {type: string}
        intimacy_level: {type: number}
        trust_level: {type: number}
        dependence_level: {type: number}
        fear_level: {type: number}
        resentment_level: {type: number}
        attraction_level: {type: number}
        rivalry_level: {type: number}
        dominance_source_over_target: {type: number}
        perceived_loyalty: {type: number}
        hidden_agenda_score: {type: number}
        recent_betrayal_score: {type: number}
        unresolved_tension_score: {type: number}
        status: {type: string}
        last_interaction_scene_id: {type: string, nullable: true}
        updated_at: {type: string, format: date-time}
    RelationshipGraph:
      type: object
      properties:
        project_id: {type: string}
        nodes:
          type: array
          items:
            type: object
            properties:
              character_id: {type: string, format: uuid}
        edges:
          type: array
          items:
            $ref: "#/components/schemas/RelationshipRead"
...
    CompileSceneDirectRequest:
      type: object
      required: [scene_context, scene_analysis]
      properties:
        scene_context:
          type: object
          additionalProperties: true
        scene_analysis:
          type: object
          additionalProperties: true
        previous_scene_state:
          type: object
          additionalProperties: true
          nullable: true
        character_arc_state:
          type: object
          additionalProperties: true
          nullable: true
...
    CompileEpisodeRequest:
      type: object
      required: [project_id, episode_id]
      properties:
        project_id: {type: string, format: uuid}
        episode_id: {type: string, format: uuid}
        force_recompute: {type: boolean, default: false}
    CompileEpisodeResponse:
      type: object
      properties:
        project_id: {type: string, format: uuid}
        episode_id: {type: string, format: uuid}
        scene_count: {type: integer}
        compiled_scenes:
          type: array
          items:
            type: object
            additionalProperties: true
        continuity_warnings:
          type: array
          items:
            type: object
            additionalProperties: true
    CompileProjectRequest:
      type: object
      required: [project_id]
      properties:
        project_id: {type: string, format: uuid}
        force_recompute: {type: boolean, default: false}
    CompileProjectResponse:
      type: object
      properties:
        project_id: {type: string, format: uuid}
        episode_count: {type: integer}
        scene_count: {type: integer}
        compiled_scenes:
          type: array
          items:
            type: object
            additionalProperties: true
        continuity_warnings:
          type: array
          items:
            type: object
            additionalProperties: true
...